web-server/requirements.txt
Django
django-leaflet==0.28.3
python-decouple==3.6
django-filter==22.1
geoip2==4.6.0